refact(discovery): move field path related helper code to a submodule
